Ks login

+++ src/components/auth/Login.js
@@ -1,5 +1,4 @@
 import React, { useRef } from "react"
-import { Link } from "react-router-dom";
 import "./Login.css"
 
 
@@ -21,14 +20,20 @@ export const Login = props => {
 
         existingUserCheck()
             .then(exists => {
-                if (exists && exists.password === password.current.value) {
+                if (exists && exists.password === password.current.value && exists.userTypeId === 1) {
                     // The user id is saved under the key app_user_id in local Storage. Change below if needed!
                     localStorage.setItem("app_user_id", exists.id)
                     props.history.push("/")
+                } else if (exists && exists.password === password.current.value && exists.userTypeId === 2) {
+                    // The user id is saved under the key app_user_id in local Storage. Change below if needed!
+                    localStorage.setItem("app_user_id", exists.id)
+                    props.history.push("/dashboards")
                 } else if (exists && exists.password !== password.current.value) {
                     passwordDialog.current.showModal()
                 } else if (!exists) {
                     existDialog.current.showModal()
+                } else {
+                    console.log("pit of despair")
                 }
             })
     }
@@ -45,7 +50,7 @@ export const Login = props => {
             </dialog>
             <section>
                 <form className="form--login" onSubmit={handleLogin}>
-                    <h1>Application Name</h1>
+                    <h1>Bespoke Hub</h1>
                     <h2>Please sign in</h2>
                     <fieldset>
                         <label htmlFor="inputEmail"> Email address </label>
@@ -64,15 +69,15 @@ export const Login = props => {
                             required />
                     </fieldset>
                     <fieldset>
-                        <button type="submit">
+                        <button type="submit" onClick={handleLogin} >
                             Sign in
                         </button>
                     </fieldset>
                 </form>
             </section>
-            <section className="link--register">
+            {/* <section className="link--register">
                 <Link to="/register">Not a member yet?</Link>
-            </section>
+            </section> */}
         </main>
     )
 }
